fix: 로그인 성공 후 시작 페이지로 돌아오는 무한루프 문제 해결
